requirements.txt
beautifulsoup4
dnspython
pyfiglet
requests
tqdm
Wappalyzer
python-whois
pyOpenSSL